+++ drake/systems/controllers/CMakeLists.txt
@@ -65,9 +65,11 @@ pods_install_pkg_config_file(drake-zmp-util
 
 # System 2 controllers
 set(system2_sources
-    gravity_compensator.cc)
+    gravity_compensator.cc
+    pid_controller.cc)
 set(system2_installed_headers
-    gravity_compensator.h)
+    gravity_compensator.h
+    pid_controller.h)
 add_library_with_exports(LIB_NAME drakeSystemControllers
     SOURCE_FILES ${system2_sources} ${system2_installed_headers})
 add_dependencies(drakeSystemControllers drakeCommon drakeRBM drakeSystemFramework)
