+++ drake/systems/controllers/test/CMakeLists.txt
@@ -4,3 +4,7 @@ target_link_libraries(gravity_compensator_test
     drakeSystemFramework
     drakeRBSystem GTest::GTest GTest::Main)
 add_test(NAME gravity_compensator_test COMMAND gravity_compensator_test)
+
+
+drake_add_cc_test(pid_controller_test)
+target_link_libraries(pid_controller_test drakeSystemControllers drakeSystemFramework)
