+++ drake/systems/framework/primitives/CMakeLists.txt
@@ -13,7 +13,6 @@ drake_install_headers(
   multiplexer.h
   pass_through.h
   pass_through-inl.h
-  pid_controller.h
   time_varying_polynomial_source.h
   zero_order_hold.h)
