+++ drake/systems/plants/controlled_spring_mass_system/CMakeLists.txt
@@ -2,11 +2,12 @@ add_library_with_exports(LIB_NAME drakeControlledSpringMassSystemPlant SOURCE_FI
     controlled_spring_mass_system.cc)
 target_link_libraries(drakeControlledSpringMassSystemPlant
     drakeSpringMassSystemPlant
+    drakeSystemControllers
     drakeSystemFramework)
 pods_install_libraries(drakeControlledSpringMassSystemPlant)
 drake_install_headers(controlled_spring_mass_system.h)
 
 pods_install_pkg_config_file(drake-controlled-spring-mass-system-plant
-    LIBS -ldrakeControlledSpringMassSystemPlant -ldrakeSpringMassSystemPlant -ldrakeSystemFramework
+    LIBS -ldrakeControlledSpringMassSystemPlant -ldrakeSpringMassSystemPlant -drakeSystemControllers -ldrakeSystemFramework
     REQUIRES
     VERSION 0.0.1)
