+++ drake/systems/plants/controlled_spring_mass_system/controlled_spring_mass_system.h
@@ -8,7 +8,7 @@
 #include "drake/systems/framework/primitives/constant_vector_source.h"
 #include "drake/systems/framework/primitives/demultiplexer.h"
 #include "drake/systems/framework/primitives/gain.h"
-#include "drake/systems/framework/primitives/pid_controller.h"
+#include "drake/systems/controllers/pid_controller.h"
 #include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
 
 namespace drake {
